chore(docker): consolidate docker code
This patch consolidates all of docker implementation under the
`docker/` folder.
